fix(iptv): o preço do plano estava em CENTAVOS, contra o banco inteiro
Escrevi `preco integer` e comentei "em centavos, 1990 = R$ 19,90". Era eu
contra o padrão de todo o resto:

  cobrancas.valor          numeric   `[{"nome":"Corte","valor":45}]`
  appointments.price       numeric   "como o resto do dinheiro deste banco"
  billing.payments.amount  numeric

E o `formatMoney`, que desenha dinheiro em toda tela do produto, não
divide por nada — ele recebe reais.

## Como apareceu

Simulando um mês de loja: o preço saiu do plano em "centavos" e entrou em
`cobrancas.valor`, que é reais. O caixa somou e mostrou **R$ 19.950** de
comissão onde eram R$ 199,50 — cem vezes, na tela que diz quanto pagar a
cada funcionário.

Ninguém tinha vendido nada ainda, então nenhuma conta de verdade saiu
errada. A primeira sairia.

Convertido dividindo por 100, que foi como os três planos do piloto
entraram: 1990, 4990, 11990 e 1500 viram 19,90, 49,90, 119,90 e 15,00.

+++ supabase/schemas/03_models/03-30_iptv.sql
@@ -128,17 +128,20 @@ create table if not exists public.iptv_pacotes (
    * testes. Quem custa é o pago: 1 no mensal, 12 no anual.
    */
   creditos integer,
-  -- Preço de venda em CENTAVOS, como todo dinheiro neste banco. 1990 = 19,90.
-  preco integer,
+  -- Preço de venda em REAIS, como todo dinheiro neste banco. 19.90 = R$ 19,90.
+  preco numeric,
   /**
-   * Centavos por tela ALÉM das que o pacote já inclui.
+   * Reais por tela ALÉM das que o pacote já inclui.
    *
    * Por plano, e não uma regra no código: quantas telas vêm juntas depende do
    * plano E do servidor. O COMPLETO do piloto traz duas, o UNITV traz uma, e o
    * próximo servidor trará outro número. Regra fixa erraria no primeiro
    * servidor novo — e erraria cobrando.
+   *
+   * Em REAIS. Nasceu em centavos, contra o padrão do banco inteiro, e o erro
+   * só apareceu quando o caixa somou comissão: R$ 19.950 onde eram R$ 199,50.
    */
-  preco_tela_extra integer,
+  preco_tela_extra numeric,
   /**
    * Renova no painel quando vencer, sem cobrar de novo. É o vitalício.
    *
